+++ r-package/NEWS.md
@@ -3,6 +3,7 @@
 **Bug fixes**
 
 - Fix bug in `read_immediate_region()` which was hardcoded to read state 11 by mistake.
+- The function `remove_islands()` now correctly drops the arquipelago de Trindade e Martim Vaz
 
 
 # geobr v2.0.0